# docs/api/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: GHARMIND AI — Household Operating System API
  description: |
    API for GHARMIND AI — India's First AI-powered Household Operating System.
    
    Powers the Digital Twin, Prediction Engine, What-If Simulator, and 
    conversational household AI (Gharji).
    
    All timestamps are in ISO 8601 format with Asia/Kolkata (IST) timezone.
    Authentication via AWS Cognito JWT.
  version: "1.0.0"
  contact:
    name: GHARMIND AI Team
    email: dev@gharmind.in
  license:
    name: Proprietary
servers:
  - url: https://api.gharmind.in/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Development
security:
  - CognitoJWT: []
...
paths:
...
  /households/{householdId}/simulator/run:
    post:
      tags: [Simulator]
      summary: Run a What-If simulation
      operationId: runSimulation
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRunRequest'
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResult'
        '422':
          $ref: '#/components/responses/ValidationError'
  /households/{householdId}/simulator/scenarios:
    get:
      tags: [Simulator]
      summary: List built-in scenario templates
      operationId: getScenarios
      parameters:
        - $ref: '#/components/parameters/HouseholdId'
      responses:
        '200':
          description: Scenario templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioList'

  # ─── CHAT ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito Access Token
  parameters:
    HouseholdId:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Household UUID
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string

    # ── Household ─────────────────────────────────────────────
...
    SimulationRunRequest:
      type: object
      required: [scenario_name, hypothesis, perturbations, sim_start, sim_duration_hours]
      properties:
        scenario_name:
          type: string
          maxLength: 200
        hypothesis:
          type: string
          description: Natural language description of what is being tested
        perturbations:
          type: array
          items:
            $ref: '#/components/schemas/Perturbation'
          minItems: 1
        sim_start:
          type: string
          format: date-time
        sim_duration_hours:
          type: integer
          minimum: 1
          maximum: 48
    Perturbation:
      type: object
      required: [type, params]
      properties:
        type:
          type: string
          enum: [power_cut, water_shortage, internet_outage, unexpected_guest, member_absent, exam_week, festival_tomorrow, motor_failure, gas_cylinder_empty]
        params:
          type: object
          additionalProperties: true
    SimulationResult:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [running, complete, failed]
        scenario_name:
          type: string
        result_summary:
          type: string
          description: Plain English summary in user's language
        overall_severity:
          type: string
          enum: [minimal, moderate, significant, critical]
        impact_analysis:
          type: object
          properties:
            disrupted_routines:
              type: array
              items:
                type: object
            resource_impacts:
              type: array
              items:
                type: object
            cascade_chain:
              type: array
              items:
                type: string
        action_plan:
          type: array
          items:
            type: object
            properties:
              time:
                type: string
              action:
                type: string
        risk_flags:
          type: array
          items:
            type: object
    ScenarioList:
      type: object
      properties:
        scenarios:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              description:
                type: string
              perturbation_types:
                type: array
                items:
                  type: string

    # ── Chat ──────────────────────────────────────────────────
...